Port of the udpate to use MavenWrapper
- Ported the ability to use the MavenWrapper in the docker image
- Replaced the integration tests with ones from main - that actually test different packagings of contracts
