+++ fabric-chaincode-integration-test/src/contracts/bare-gradle/src/main/resources/config.props
