+++ fabric-chaincode-integration-test/src/contracts/bare-maven/.gitignore
@@ -1,0 +1,1 @@
+target
