--- fabric-chaincode-integration-test/src/contracts/fabric-chaincode-example-sacc/build.gradle
@@ -1,29 +1,0 @@
-plugins {
-    id 'com.github.johnrengelman.shadow' version '5.1.0'
-    id 'java'
-}
-
-group 'org.hyperledger.fabric-chaincode-java'
-version '1.0-SNAPSHOT'
-
-sourceCompatibility = 1.8
-
-repositories {
-    mavenLocal()
-    mavenCentral()
-}
-
-dependencies {
-    compile group: 'org.hyperledger.fabric-chaincode-java', name: 'fabric-chaincode-shim', version: '2.2.2'
-    testCompile group: 'junit', name: 'junit', version: '4.12'
-}
-
-shadowJar {
-    baseName = 'chaincode'
-    version = null
-    classifier = null
-
-    manifest {
-        attributes 'Main-Class': 'org.hyperledger.fabric.contract.ContractRouter'
-    }
-}
