--- fabric-chaincode-integration-test/src/contracts/fabric-chaincode-example-sbe/settings.gradle
@@ -1,2 +1,0 @@
-rootProject.name = 'fabric-chaincode-example-sbe'
-
