+++ fabric-chaincode-integration-test/src/contracts/fabric-shim-api/build.gradle
@@ -1,0 +1,34 @@
+plugins {
+    id 'com.github.johnrengelman.shadow' version '5.1.0'
+    id 'java'
+}
+
+group 'org.hyperledger.fabric-chaincode-java'
+version '1.0-SNAPSHOT'
+
+sourceCompatibility = 1.8
+
+repositories {
+    mavenLocal()
+    mavenCentral()
+    maven { url = "https://www.jitpack.io" }
+}
+
+dependencies {
+    implementation group: 'org.hyperledger.fabric-chaincode-java', name: 'fabric-chaincode-shim', version: '2.2.2'
+    implementation group: 'org.hyperledger.fabric-chaincode-java', name: 'fabric-chaincode-protos', version: '2.2.2'
+    testImplementation group: 'junit', name: 'junit', version: '4.12'
+    implementation group: 'commons-logging', name: 'commons-logging', version: '1.2'
+    implementation group: 'com.google.code.gson', name: 'gson', version: '2.8.6'
+    implementation group: 'com.google.protobuf', name: 'protobuf-java', version: '3.4.0'
+}
+
+shadowJar {
+    baseName = 'chaincode'
+    version = null
+    classifier = null
+
+    manifest {
+        attributes 'Main-Class': 'org.hyperledger.fabric.contract.ContractRouter'
+    }
+}
