+++ fabric-chaincode-integration-test/src/contracts/fabric-shim-api/settings.gradle
