+++ fabric-chaincode-integration-test/src/contracts/fabric-shim-api/src/main/resources/config.props
@@ -1,0 +1,6 @@
+MAX_INBOUND_MESSAGE_SIZE=4000
+CHAINCODE_METRICS_ENABLED=true
+TP_CORE_POOL_SIZE=4
+TP_MAX_POOL_SIZE=4
+TP_QUEUE_SIZE=4000
+
