+++ fabric-chaincode-integration-test/src/contracts/wrapper-maven/.gitignore
@@ -1,0 +1,1 @@
+target
