+++ fabric-chaincode-integration-test/src/test/resources/first-network/scripts/script.sh
@@ -7,8 +7,7 @@ echo "\___ \    | |     / _ \   | |_) |   | |  "
 echo " ___) |   | |    / ___ \  |  _ <    | |  "
 echo "|____/    |_|   /_/   \_\ |_| \_\   |_|  "
 echo
-echo "Build your first network (BYFN) end-to-end test"
-echo
+
 CHANNEL_NAME="$1"
 DELAY="$2"
 LANGUAGE="$3"
@@ -23,15 +22,6 @@ LANGUAGE=`echo "$LANGUAGE" | tr [:upper:] [:lower:]`
 COUNTER=1
 MAX_RETRY=10
 echo "Channel name : "$CHANNEL_NAME
-
-if [ "$CHANNEL_NAME" = "sachannel" ]; then
-  CC_SRC_PATH="/opt/gopath/src/github.com/hyperledger/fabric/peer/chaincodes/sacc"
-  CC_NAME="javacc"
-elif [ "$CHANNEL_NAME" = "sbechannel" ]; then
-  CC_SRC_PATH="/opt/gopath/src/github.com/hyperledger/fabric/peer/chaincodes/sbe"
-  CC_NAME="sbecc"
-  COLLECTIONS_CFG=$(realpath scripts/collection_config.json)
-fi
 
 # import utils
 . scripts/utils.sh
@@ -50,12 +40,55 @@ updateAnchorPeers 0 1
 echo "Updating anchor peers for org2..."
 updateAnchorPeers 0 2
 
+## There are several chaincodes that need packaging and installing
+## Currently using the old lifecyle
+CC_ROOT_PATH_IN_DOCKER="/opt/gopath/src/github.com/hyperledger/fabric/peer/chaincodes"
+CC_SRC_PATH="${CC_ROOT_PATH_IN_DOCKER}/fabric-shim-api"
+CC_NAME="shimcc"
+COLLECTIONS_CFG=$(realpath scripts/collection_config.json)
+
 echo "Installing chaincode on peer 0, org 1"
 installChaincode 0 1
 echo "Installing chaincode on peer 0, org 2"
 installChaincode 0 2
-
 echo "Instantiating chaincode on peer 0, org 1"
 instantiateChaincode 0 1
 
+#
+CC_SRC_PATH="${CC_ROOT_PATH_IN_DOCKER}/bare-gradle"
+CC_NAME="baregradlecc"
+unset COLLECTIONS_CFG
+
+echo "Installing chaincode on peer 0, org 1"
+installChaincode 0 1
+echo "Installing chaincode on peer 0, org 2"
+installChaincode 0 2
+echo "Instantiating chaincode"
+instantiateChaincode 0 1
+
+#
+CC_SRC_PATH="${CC_ROOT_PATH_IN_DOCKER}/bare-maven"
+CC_NAME="baremaven"
+unset COLLECTIONS_CFG
+
+echo "Installing chaincode on peer 0, org 1"
+installChaincode 0 1
+echo "Installing chaincode on peer 0, org 2"
+installChaincode 0 2
+echo "Instantiating chaincode"
+instantiateChaincode 0 1
+
+#
+CC_SRC_PATH="${CC_ROOT_PATH_IN_DOCKER}/wrapper-maven"
+CC_NAME="wrappermaven"
+unset COLLECTIONS_CFG
+
+echo "Installing chaincode on peer 0, org 1"
+installChaincode 0 1
+echo "Installing chaincode on peer 0, org 2"
+installChaincode 0 2
+echo "Instantiating chaincode"
+instantiateChaincode 0 1
+
+echo "<< DONE"
 # exit 0
